requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
python-docx>=1.1.0
python-pptx>=1.0.0
pypdf>=4.0.0
fpdf2>=2.7.0
lxml>=5.0.0
openpyxl>=3.1.0
msal>=1.28.0
requests>=2.31.0
